OneDrive/Documents/BigData/marino-mongodb/pyproject.toml
[project]
name = "marino-mongodb"
version = "0.1.0"
description = ""
authors = [
    {name = "Ella Marino",email = "marino77@students.rowan.edu"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymongo (>=4.12.1,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
